Feat: input state에 따라 채팅 전송 버튼과 일기 생성 버튼을 조건부 렌더링 (#231)
