+++ src/features/chatting/containers/ChatContainer.tsx
@@ -1,22 +1,26 @@
 import {
   getGetCalendarViewQueryKey,
   getGetCurrentStreakQueryKey,
+  getGetFeedQueryKey,
   getGetMessagesQueryKey,
   useCreateChatDiary,
   useGetMessages,
   useMessageToAiTest,
 } from '@/src/apis/_generated/serverAPI';
 import { ApiResponseChatListResponse } from '@/src/apis/_generated/serverAPI.schemas';
-import { COLOR } from '@/src/constants/theme';
+import { COLOR, FONT_FAMILY } from '@/src/constants/theme';
+import { useIsKeyboardOpen } from '@/src/hooks/useIsKeyboardOpen';
 import { toast } from '@/src/modules/toast';
+import responsiveToPx from '@/src/utils/responsiveToPx';
 import { useQueryClient } from '@tanstack/react-query';
 import { useRouter } from 'expo-router';
+import { useState } from 'react';
 import { KeyboardAvoidingView } from 'react-native-keyboard-controller';
 import { SafeAreaView } from 'react-native-safe-area-context';
 import styled from 'styled-components/native';
 import ChatHeader from '../components/ChatHeader';
-import ChatInput from '../components/ChatInput';
 import ChatScrollView from '../components/ChatScrollView';
+import ChatSendButton from '../components/ChatSendButton';
 import GenerateDiaryButton from '../components/GenerateDiaryButton';
 import { useCanGenerateDiary } from '../hooks/useCanGenerateDiary';
 import { useChattingQueryParams } from '../hooks/useChattingQueryParams';
@@ -24,7 +28,9 @@ import { chatListFilter } from '../utils';
 import { setOptimisticUserMessage } from '../utils/optimisticUpdate';
 
 const ChatContainer = () => {
+  const [input, setInput] = useState<string>('');
   const { aiProfileId, year, month, day } = useChattingQueryParams();
+  const isKeyboardOpen = useIsKeyboardOpen();
   const queryClient = useQueryClient();
   const router = useRouter();
 
@@ -63,6 +69,7 @@ const ChatContainer = () => {
     mutation: {
       onSettled: () => {
         queryClient.invalidateQueries({ queryKey: getGetCalendarViewQueryKey({ year, month }) });
+        queryClient.invalidateQueries({ queryKey: getGetFeedQueryKey() });
         queryClient.invalidateQueries({ queryKey: getGetCurrentStreakQueryKey() });
       },
       onSuccess: () => {
@@ -73,6 +80,9 @@ const ChatContainer = () => {
         if (e.response?.status === 429) {
           return toast({ message: '사용 한도를 초과했습니다.', options: { type: 'error' } });
         }
+        if (e.response?.data.code === 'DIARY4004') {
+          return toast({ message: '일기는 하루 3개까지 작성 가능합니다.', options: { type: 'error' } });
+        }
         toast({ message: '일기 생성에 실패했습니다.', options: { type: 'error' } });
       },
     },
@@ -82,9 +92,12 @@ const ChatContainer = () => {
     router.back();
   };
 
-  const handleInputSubmit = (content: string) => {
+  const handleInputSubmit = () => {
     if (sendChatMutation.isPending) return;
+    const content = input.trim();
+    if (!content) return;
     sendChatMutation.mutate({ data: { aiProfileId, year, month, day, content } });
+    setInput('');
   };
 
   const handleDiarySummaryClick = () => {
@@ -102,8 +115,24 @@ const ChatContainer = () => {
       <ChatHeader characterId={aiProfileId} onBackClick={handleBackClick} />
       <KeyboardAvoidingView style={{ flex: 1 }} behavior="padding">
         <ChatScrollView chatData={chatList} characterId={aiProfileId} year={year} month={month} day={day} />
-        <GenerateDiaryButton isLoading={generateDiaryMutation.isPending} onClick={handleDiarySummaryClick} />
-        <ChatInput isLoading={sendChatMutation.isPending} onInputSubmit={handleInputSubmit} />
+        <InputBarWrapper>
+          <StyledInput
+            value={input}
+            onChangeText={setInput}
+            multiline
+            placeholder="오늘 하루에 대해 말해주세요."
+            placeholderTextColor={COLOR.placeholder}
+          />
+          {!!input ? (
+            <ChatSendButton isLoading={sendChatMutation.isPending} onClick={handleInputSubmit} />
+          ) : (
+            <GenerateDiaryButton
+              isLoading={generateDiaryMutation.isPending}
+              isExpanded={!isKeyboardOpen}
+              onClick={handleDiarySummaryClick}
+            />
+          )}
+        </InputBarWrapper>
       </KeyboardAvoidingView>
     </SafeView>
   );
@@ -115,4 +144,23 @@ const SafeView = styled(SafeAreaView)`
   flex: 1;
   background-color: ${COLOR.background};
   padding: 0 18px;
+`;
+
+const InputBarWrapper = styled.View`
+  flex-direction: row;
+  width: 100%;
+  height: ${responsiveToPx('72px')};
+  align-items: center;
+  gap: 8px;
+  padding: 10px 0;
+`;
+
+const StyledInput = styled.TextInput`
+  flex: 1;
+  border-radius: 99px;
+  padding: 15px;
+  background-color: ${COLOR.white};
+  font-family: ${FONT_FAMILY.pretendard500};
+  font-size: 15px;
+  color: ${COLOR.title};
 `;
